requirements.txt
discord.py==2.1.0a4632+gcd6fd13a
async-timeout==3.0.1
dnspython==1.16.0
PyNaCl==1.3.0
google-auth==2.6.6
google-auth-oauthlib==0.5.1
gspread==5.3.2
python-dotenv==0.20.0
pymongo==4.1.1